JoinRoom のモーダルをコンポーネントに切り出した
