+++ react/src/features/top/components/JoinRoom/JoinRoom.tsx
@@ -1,7 +1,9 @@
 import { useState, type FormEventHandler } from 'react';
 import { joinRoom } from '../../api';
 import { JoinRoomButton } from './JoinRoomButton';
+import { JoinRoomModal } from './JoinRoomModal';
 import { ExhaustiveError } from '@/error';
+import { useModal } from '@/hooks/useModal';
 
 type Props = {
   className: string;
@@ -10,14 +12,7 @@ type Props = {
 export const JoinRoom: React.FC<Props> = ({ className }) => {
   const [roomId, setRoomId] = useState('');
   const [joinRoomResult, setJoinRoomResult] = useState('');
-  const [open, setOpen] = useState(false);
-
-  const onOpenModal = () => {
-    setOpen(true);
-  };
-  const onCloseModal = () => {
-    setOpen(false);
-  };
+  const { open, onOpenModal, onCloseModal } = useModal();
 
   const handleSubmit: FormEventHandler<HTMLFormElement> = async (event) => {
     event.preventDefault();
@@ -46,15 +41,16 @@ export const JoinRoom: React.FC<Props> = ({ className }) => {
   };
 
   return (
-    <JoinRoomButton
-      className={className}
-      open={open}
-      onOpenModal={onOpenModal}
-      onCloseModal={onCloseModal}
-      roomId={roomId}
-      setRoomId={setRoomId}
-      joinRoomResult={joinRoomResult}
-      handleSubmit={handleSubmit}
-    />
+    <>
+      <JoinRoomButton className={className} onOpenModal={onOpenModal} />
+      <JoinRoomModal
+        open={open}
+        onCloseModal={onCloseModal}
+        roomId={roomId}
+        setRoomId={setRoomId}
+        joinRoomResult={joinRoomResult}
+        handleSubmit={handleSubmit}
+      />
+    </>
   );
 };
